both parts, this needs cleanup
thanks Reddit. I had figured out the concept of inclusion-exclusion but could not figure out
how to implement it properly. but it is actually quite simple: if we are intersecting a positive
cuboid, the intersection should be negative, and vice versa. it does not matter what the attacker is.
maybe the original messed up implementation would lead to something but playing around with working
code from someone else I was able to figure out I had an error in `no_overlap` function - I was
checking for `all` instead of `any` which is why I ended with more intersections than needed.