[day19] gen. of rotated coordinates, hear me out
This generator creates some overhead because we yield after each new set
of coordinates is generated (actually in this moment it is a list of
coordinates). We keep track of already generated configurations and use
those to generate further configurations. We first roll, then pitch, and
finally yaw.

Although if potentially left to yield until last configuration this
approach generates redundant configurations (worst case scenario we'd
yield 84 configurations, while only 24 are unique), I think it will
rarely (if ever) happen simply because we are going to be checking the
generated configurations for the "12 matching coordinates" rule and at
that point stop yielding.

My hunch is that optimizing this generator to yield only unique configs
is not worth it, or that it could even eat up more time.

NEXT STEP: implement the 12 point check between two sets of coordinates:
- sort both lists
- generate some sort of map of distances/differences for every coord in
each list
- try to find the "pattern" in the other list
- it pattern matches for 12 or more - return True

I still need to think about this